+++ docs/changes.rst
@@ -11,6 +11,7 @@ Weblate 5.12
 
 * :ref:`check-same` is now skipped for Toki Pona.
 * :ref:`search-strings` supports ``has:location`` lookup.
+* Unicode whitespaces are now considered as problematic characters for :ref:`glossary`, see :ref:`check-prohibited-initial-character`.
 
 .. rubric:: Bug fixes